pyproject.toml
[tool.poetry]
name = "pick-dinner-for-me"
version = "0.1.0"
description = ""
authors = ["Sheng Yunzhou <shengyunzhou@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
